+++ packages/tree-view/spec/tree-view-package-spec.js
@@ -3043,7 +3043,10 @@ describe("TreeView", function () {
         fileView.dispatchEvent(new MouseEvent('click', {bubbles: true, detail: 1}));
         treeView.focus();
 
-        spyOn(shell, 'moveItemToTrash').andReturn(false);
+        spyOn(shell, 'trashItem').andCallFake(() => {
+          return Promise.reject(false);
+        })
+
         spyOn(atom, 'confirm').andCallFake((options, callback) => callback(0));
 
         atom.commands.dispatch(treeView.element, 'tree-view:remove');
@@ -3094,7 +3097,9 @@ describe("TreeView", function () {
             return atom.commands.dispatch(treeView.element, 'tree-view:remove');
           });
 
-          waitsFor('directory to be deleted', () => callback.mostRecentCall.args[0].pathToDelete === dirPath2);
+          waitsFor('directory to be deleted', () =>
+            callback.mostRecentCall?.args?.[0].pathToDelete === dirPath2
+          );
 
           return runs(function () {
             const openFilePaths = atom.workspace.getTextEditors().map(editor => editor.getPath());
@@ -3123,7 +3128,9 @@ describe("TreeView", function () {
             return atom.commands.dispatch(treeView.element, 'tree-view:remove');
           });
 
-          waitsFor('directory to be deleted', () => callback.mostRecentCall.args[0].pathToDelete === dirPath2);
+          waitsFor('directory to be deleted', () =>
+            callback.mostRecentCall?.args?.[0].pathToDelete === dirPath2
+          );
 
           return runs(function () {
             const openFilePaths = atom.workspace.getTextEditors().map(editor => editor.getPath());
@@ -3157,7 +3164,9 @@ describe("TreeView", function () {
             return atom.commands.dispatch(treeView.element, 'tree-view:remove');
           });
 
-          waitsFor('directory to be deleted', () => callback.mostRecentCall.args[0].pathToDelete === dirPath2);
+          waitsFor('directory to be deleted', () =>
+            callback.mostRecentCall?.args?.[0].pathToDelete === dirPath2
+          );
 
           return runs(function () {
             const openFilePaths = atom.workspace.getTextEditors().map(editor => editor.getPath());
@@ -3188,7 +3197,9 @@ describe("TreeView", function () {
             return atom.commands.dispatch(treeView.element, 'tree-view:remove');
           });
 
-          waitsFor('directory to be deleted', () => callback.mostRecentCall.args[0].pathToDelete === dirPath2);
+          waitsFor('directory to be deleted', () =>
+            callback.mostRecentCall?.args?.[0].pathToDelete === dirPath2
+          );
 
           return runs(function () {
             const openFilePaths = atom.workspace.getTextEditors().map(editor => editor.getPath());
@@ -3196,10 +3207,14 @@ describe("TreeView", function () {
           });
         });
 
+<<<<<<< HEAD
         it("focuses the directory's parent folder", function () {
+=======
+        it("focuses the directory's parent folder", async function () {
+          jasmine.useRealClock();
+>>>>>>> bce72212de5f12f4846b065f9379151dc23fc515
           const callback = jasmine.createSpy('onEntryDeleted');
           treeView.onEntryDeleted(callback);
-
           dirView2.dispatchEvent(new MouseEvent('click', {bubbles: true, detail: 1}));
           treeView.focus();
 
@@ -3207,11 +3222,17 @@ describe("TreeView", function () {
 
           atom.commands.dispatch(treeView.element, 'tree-view:remove');
 
-          waitsFor('directory to be deleted', () => callback.mostRecentCall.args[0].pathToDelete === dirPath2);
+          waitsFor('directory to be deleted', () =>
+            callback.mostRecentCall?.args?.[0].pathToDelete === dirPath2
+          );
 
           runs(() => {
+<<<<<<< HEAD
             console.log('most recent', callback.mostRecentCall);
             expect(root1).toHaveClass('selected')
+=======
+            expect(root1).toHaveClass('selected');
+>>>>>>> bce72212de5f12f4846b065f9379151dc23fc515
           });
         });
       });
@@ -3765,14 +3786,20 @@ describe("TreeView", function () {
         treeView.onEntryDeleted(callback);
 
         const pathToDelete = treeView.selectedEntry().getPath();
-        expect(treeView.selectedEntry().getPath()).toContain(path.join('dir2', 'new2'));
+        expect(
+          treeView.selectedEntry().getPath()
+        ).toContain(path.join('dir2', 'new2'));
+
         const dirView = findDirectoryContainingText(treeView.roots[0], 'dir2');
         expect(dirView).not.toBeNull();
         spyOn(dirView.directory, 'updateStatus');
+
         spyOn(atom, 'confirm').andCallFake((options, callback) => callback(0));
         atom.commands.dispatch(treeView.element, 'tree-view:remove');
 
-        waitsFor('onEntryDeleted to be called', () => callback.mostRecentCall.args[0].pathToDelete === pathToDelete);
+        waitsFor('onEntryDeleted to be called', () =>
+          callback.mostRecentCall?.args?.[0].pathToDelete === pathToDelete
+        );
 
         return runs(() => expect(dirView.directory.updateStatus).toHaveBeenCalled());
       }));